+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docstring parsing & failure prevention
 - Description of class and attributes in jsonschema
 
+### Changed
+- Union cannot implement two classes with non coherent standalone_in_db attributes anymore
+
+### Removed
+- TypedDict not supported anymore
+
 ## [0.4.3]
 ### Added
 - Datatype from jsonschema method